Move uefi_fv to library.uefi
