+++ chipsec/utilcmd/uefi_cmd.py
@@ -65,8 +65,8 @@ from chipsec.hal.common.spi_uefi import decode_uefi_region, modify_uefi_region, 
 from chipsec.hal.common.spi_uefi import CMD_UEFI_FILE_INSERT_AFTER, CMD_UEFI_FILE_INSERT_BEFORE, CMD_UEFI_FILE_REMOVE
 from chipsec.hal.common.uefi import UEFI, decode_EFI_variables, get_attr_string, identify_EFI_NVRAM
 from chipsec.hal.common.uefi import SECURE_BOOT_KEY_VARIABLES, parse_script, parse_EFI_variables
-from chipsec.hal.common.uefi_fv import get_guid_bin, assemble_uefi_file, assemble_uefi_section, assemble_uefi_raw
-from chipsec.hal.common.uefi_fv import FILE_TYPE_NAMES
+from chipsec.library.uefi.uefi_fv import get_guid_bin, assemble_uefi_file, assemble_uefi_section, assemble_uefi_raw
+from chipsec.library.uefi.uefi_fv import FILE_TYPE_NAMES
 from chipsec.hal.common.uefi_platform import fw_types
